cleanup pytest configuration
